requirements.txt
astroid==2.3.3
autopep8==1.5
blessings==1.7
certifi==2019.11.28
chardet==3.0.4
Django==2.2.13
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.8
inquirer==2.6.3
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.3
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-editor==1.0.4
pytz==2019.3
readchar==2.0.1
requests==2.22.0
requests-cache==0.5.2
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
tqdm==4.42.1
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
xmltodict==0.12.0
